Update README.md
The universe is still a mystery for humans, for thousands of years our predecessors have looked at the stars and wondered what could be out there. In the last 60 years, our knowledge about the universe has increased at an incredible rate, but there are still hundreds of questions yet to be answered. The main focus of this challenge is to portray an image of what we imagine a life-sustaining exoplanet could look like based on what we know so far about other planets and astrological conditions. With this challenge, we will envision what it would be like to explore one of these planets and be able to experience the wonders of nature. With the 3D rendering of our exoplanet, it will give every user the opportunity to visit another planet that can sustain life. The user will be able to visit several different locations to experience the natural attractions our exoplanet has to offer. The user will live beyond the unreal experience, something in the boundary between science and fiction. We consider it important since first of all it will encourage and show others the possibilities of life outside of planet Earth, even a better life. Our planet is suffering a lot from contamination, climate change, and global warming. What if in the future it becomes inhabitable? Where do we go? Well, that´s when the exoplanets enter the conversation. Also what if we can have a better quality of life, there are endless possibilities... but we don´t know if Earth will always be the primary source of life, and there´s only one way to avoid a catastrophe in the future, have the fundamentals, research, sources, support and students like us to make life a nice and stable journey.